moving log subscriber to rails/rack
